backend/requirements.txt
Flask
Flask-Cors
psycopg2-binary
requests
xlrd
ddddocr
